add scripts; add .gitignore
