+++ conf/conf.go
@@ -13,12 +13,14 @@ var (
 )
 
 func init() {
-	Cfg.Parse()
+	if !Cfg.Parse() {
+		panic("config init error!")
+	}
 }
 
 type Mode struct {
-	Standalone bool   `mapstructure:"standalone"`
-	Signal     string `mapstructure:"signal"`
+	Single bool   `mapstructure:"single"`
+	Pprof  string `mapstructure:"pprof"`
 }
 
 type Log struct {
@@ -29,15 +31,7 @@ type Etcd struct {
 	Servers []string `mapstructure:"servers"`
 }
 
-type Centrifugo struct {
-	Url     string `mapstructure:"url"`
-	Key     string `mapstructure:"key"`
-	CertPem string `mapstructure:"certpem"`
-	KeyPem  string `mapstructure:"keypem"`
-	Expire  int64  `mapstructure:"expire"`
-}
-
-type Protoo struct {
+type Signal struct {
 	Host    string `mapstructure:"host"`
 	Port    string `mapstructure:"port"`
 	CertPem string `mapstructure:"certpem"`
@@ -49,18 +43,17 @@ type SFU struct {
 }
 
 type Config struct {
-	Mode       Mode       `mapstructure:"mode"`
-	Log        Log        `mapstructure:"log"`
-	Etcd       Etcd       `mapstructure:"etcds"`
-	Centrifugo Centrifugo `mapstructure:"centrifugo"`
-	Protoo     Protoo     `mapstructure:"protoo"`
-	Sfu        SFU        `mapstructure:"sfu"`
-	CfgFile    string
-	err        error
+	Mode    Mode   `mapstructure:"mode"`
+	Log     Log    `mapstructure:"log"`
+	Etcd    Etcd   `mapstructure:"etcds"`
+	Signal  Signal `mapstructure:"signal"`
+	Sfu     SFU    `mapstructure:"sfu"`
+	CfgFile string
+	err     error
 }
 
 func ShowHelp() {
-	fmt.Sprintf("Usage:%s {params}\n", os.Args[0])
+	fmt.Printf("Usage:%s {params}\n", os.Args[0])
 	fmt.Println("      -c {config file}")
 	fmt.Println("      -h (show help info)")
 }
@@ -78,12 +71,12 @@ func (c *Config) Load() bool {
 
 	c.err = viper.ReadInConfig()
 	if c.err != nil {
-		fmt.Errorf("config file %s read failed. %v", c.CfgFile, c.err)
+		fmt.Printf("config file %s read failed. %v\n", c.CfgFile, c.err)
 		return false
 	}
 	c.err = viper.GetViper().UnmarshalExact(c)
 	if c.err != nil {
-		fmt.Errorf("config file %s loaded failed. %v", c.CfgFile, c.err)
+		fmt.Printf("config file %s loaded failed. %v\n", c.CfgFile, c.err)
 		return false
 	}
 
@@ -95,8 +88,12 @@ func (c *Config) Parse() bool {
 	flag.StringVar(&c.CfgFile, "c", "conf/conf.toml", "config file")
 	help := flag.Bool("h", false, "help info")
 	flag.Parse()
+	if !c.Load() {
+		ShowHelp()
+		return false
+	}
 
-	if !c.Load() || *help {
+	if *help {
 		ShowHelp()
 		return false
 	}
